feat(02.1-04): add ActiveTab enum, TAB_NAMES/TAB_LABELS/TAB_ICONS constants, tab state fields
- Add ActiveTab enum (Details/Top/Extensions/Duplicates/Errors = 0..4) with as_str/from_str/from_index/from_command_id
- Add TAB_NAMES, TAB_LABELS, TAB_ICONS constants as single source of truth
- Repurpose DesktopState.active_tab: usize → ActiveTab (default Details)
- Add tab_strip, tab_panels, tab_rects, icon_font fields to DesktopState
- Add tab command IDs (3020-3024) and IDC_HAND/WM_SETCURSOR constants to ffi.rs
- Fix existing active_tab consumers in mod.rs and paint.rs (as usize cast)
- active_tab_enum unit test passes

+++ src/desktop/paint.rs
@@ -130,7 +130,7 @@ pub(super) unsafe fn draw_toolbar_background(hdc: Hdc, client: Rect, state: &Des
     SetBkMode(hdc, TRANSPARENT);
 
     for (i, &(name, left, right)) in tabs.iter().enumerate() {
-        let is_active = state.active_tab == i;
+        let is_active = state.active_tab as usize == i;
         let tab_rect = Rect {
             left,
             top: 0,
